+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- Workspace demo documentation
+- add `biggy-autocomplete-wrapper` handle.
 
 ## [0.5.0] - 2020-01-13