+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - add `biggy-autocomplete-wrapper` handle.
 
+### Fixed
+
+- Use `product.link` instead of `slugifyUrl`.
+
 ## [0.5.0] - 2020-01-13
 
 ### Changed